docs : 우승자 출력 기능 구현
