+++ index.html
@@ -24,7 +24,7 @@
       </div>
       <h4>시도할 횟수를 입력해주세요.</h4>
       <div id="racing-count-section">
-        <input type="number" id="racing-count-input" />
+        <input type="number" id="racing-count-input" disabled/>
         <button id="racing-count-submit" disabled>확인</button>
       </div>
       <h4>📄 실행 결과</h4>
