+++ src/gameStart.js
@@ -2,8 +2,7 @@ import { getCarNames } from './getCarNames.js'
 import { CAR } from './constant.js'
 import { checkRacingCount } from './checkRacingCount.js'
 import { disabledElement } from './setAttribute.js'
-
-
+import { getGameResult } from './getGameResult.js'
 
 //각 자동차 진행상황 프린트
 function printEachCarProgress(CARS){
@@ -13,16 +12,14 @@ function printEachCarProgress(CARS){
         if(randomNumber >= 4){
             car.PLUS_FORWARD();
         }
-        document.querySelector('body').innerHTML += `<p class="gameProgress">${car.name} ${car.gameProgress}</p>`
+        document.querySelector('body').innerHTML += `<p class="gameProgress">${car.name} : ${car.gameProgress}</p>`
     }
 }
 
 //게임진행 보여주기
 function printGameProgress(CARS, $racingCountInput){
     for(let i=0; i<$racingCountInput.value; i++){
-
         printEachCarProgress(CARS)
-
         document.querySelector('body').innerHTML += '<br>'
     }
 }
@@ -43,31 +40,8 @@ export function gameStart(){
     const carNames = getCarNames($carNameInput)
     const CARS = carNames.map(name => new CAR(name))
 
-    console.log(CARS)
-
-
     printGameProgress(CARS, $racingCountInput)
 
-
-    //승자 구하기
-    function getWinner(CARS){
-        let max = 0
-        //최댓값 구하기
-        for(const car of CARS){
-            if(max < car.forward){
-                max = car.forward
-            }
-        }
-        //승자 구하기
-        let winner = []
-        for(const car of CARS){
-            if(car.forward === max){
-                winner.push(car.name)
-            }
-        }
-        return winner
-    }
-    let winner = getWinner(CARS)
-    console.log(winner);
+    getGameResult(CARS)
 
 }
